:tada: v3実装

+++ model/menu.go
@@ -15,12 +15,12 @@ func (r *MatsuyaMenu) Marshal() ([]byte, error) {
 type MatsuyaMenu struct {
 	Name           string  `json:"name"`
 	Type           string  `json:"type"`
-	Price          int64   `json:"price"`
-	Calorie        int64   `json:"calorie"`
+	Price          int     `json:"price"`
+	Calorie        int     `json:"calorie"`
 	Protein        float64 `json:"protein"`
 	Lipid          float64 `json:"lipid"`
-	Carbohydrate   int64   `json:"carbohydrate"`
-	Sodium         int64   `json:"sodium"`
+	Carbohydrate   float64 `json:"carbohydrate"`
+	Sodium         int     `json:"sodium"`
 	SaltEquivalent float64 `json:"saltEquivalent"`
 	Description    string  `json:"description"`
 	ImageURL       string  `json:"imageURL"`
